ml/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
torch>=2.0.0
shap>=0.44.0
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
optuna>=3.4.0
pyarrow>=14.0.0
matplotlib>=3.7.0
plotly>=5.18.0
imbalanced-learn>=0.11.0
scipy>=1.11.0
joblib>=1.3.0
pyyaml>=6.0
